Reference

JCheck.Quickcheck — Type

Quickcheck

Contain a set of property to check through the generation of random input.

Fields

description::AbstractString: description for the instance.

rng::AbstractRNG: PRNG used to generate inputs.

predicates::PredsAssoc: predicates to check.

variables::ArgsDict: Arguments used by the predicates.

n::Int: Number of random inputs to generate.

serialize_fails::Bool: If true, serialize failing inputs to a JLSO file.

JCheck.Quickcheck — Method

Quickcheck(desc; rng=GLOBAL_RNG, n=100, serialize_fails=true)

Constructor for type Quickcheck.

Arguments

desc::AbstractString: description for the instance.

rng::AbstractRNG: PRNG used to generate inputs.

n::Int: Number of random inputs to generate.

serialize_fails::Bool: If true, serialize failing inputs to a JLSO file.

Examples

julia> qc = Quickcheck("A Test")
A Test: 0 predicate and 0 free variable.

JCheck.@add_predicate — Macro

@add_predicate qc desc pred

Add the predicate pred to the set of tests qc with description desc.

Arguments

qc: Object of type Quickcheck.

desc: String describing the predicate.

pred: Predicate in the form of an anonymous function.

Notes

The form of pred is very strict:

It has to be an anonymous function. Formally, it should be an Expr of type ->.

The type of each argument appearing on the left-hand side of pred has to be specified with the x::Type syntax.

The names of the arguments of pred matter! Specifically, in a given Quickcheck object, the type of every argument must be consistent across predicates (see examples).

Each predicate stored in a given Quickcheck object must be given a distinct description.

Examples

julia> qc = Quickcheck("A Test")
A Test: 0 predicate and 0 free variable.

julia> @add_predicate qc "Identity" (x::Float64 -> x == x)
A Test: 1 predicate and 1 free variable.
x::Float64

julia> @add_predicate qc "Sum commute" ((n::Int, x::Float64) -> n + x == x + n)
A Test: 2 predicates and 2 free variables:
n::Int64
x::Float64

julia> @add_predicate qc "Is odd" isodd(x)
ERROR: Predicate declaration must have the form of an anonymous function (... -> ...)
[...]

julia> @add_predicate qc "Is odd" (x::Int -> is_odd(x))
ERROR: A declaration for variable x already exists with type Float64; please choose another name for x
[...]

JCheck.@quickcheck — Macro

@quickcheck qc

Check the properties specified in object qc of type Quickcheck.

If qc.serialize_fails is true, serialize the failing cases to JCheck_yyyy-mm-dd_HH-MM-SS.jchk. Those can latter be analyzed using load and @getcases.